# Conflicts:
#	src/main/java/uk/gov/hmcts/opal/service/report/ReportType.java
#	src/test/java/uk/gov/hmcts/opal/service/report/ReportIdTest.java

+++ src/main/java/uk/gov/hmcts/opal/service/report/CashListReportData.java
@@ -20,8 +20,8 @@ public class CashListReportData implements ReportDataInterface {
 
     @Override
     @JsonIgnore
-    public short getNumberOfRecords() {
-        return (short) (entries == null ? 0 : entries.size());
+    public long getNumberOfRecords() {
+        return entries == null ? 0 : entries.size();
     }
 
     @Override
